Remove observers gem
