+++ spec/support/sync_delayed_job_observer.rb
@@ -1,0 +1,41 @@
+require 'delayed_job'
+
+class SyncDelayedJobObserver < Delayed::Plugin
+
+  class << self
+    attr_accessor :total_processed
+
+    def enabled?
+      @enabled
+    end
+
+    def queueing?
+      @collect
+    end
+
+    def collect!
+      @collect = true
+      @enabled = false
+      Delayed::Worker.delay_jobs = true
+    end
+
+    def enable!
+      @collect = false
+      @enabled = true
+      Delayed::Worker.delay_jobs = false
+    end
+
+    def reset!
+      @collect = true
+      @total_processed = 0
+      @enabled = false
+
+      Delayed::Worker.delay_jobs = true
+    end
+
+    def process_queue!
+      success, failure = Delayed::Worker.new(:quiet => false).work_off
+      @total_processed = success + failure
+    end
+  end
+end
